requirements.txt
streamlit==1.28.1
streamlit-webrtc==0.47.1

# State Machine & Logic
transitions==0.9.0
pydantic==2.5.0

# Voice & Audio Processing (Acoustic Biomarkers)
librosa==0.10.0
numpy==1.24.3
scipy==1.11.4
soundfile==0.12.1

# Encryption & Security
cryptography==41.0.7
pycryptodome==3.19.0

# API Integrations
requests==2.31.0
python-dotenv==1.0.0

# Database & Data
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Utilities
pandas==2.1.3
python-dateutil==2.8.2
